backend/requirements.txt
fastapi
uvicorn
SQLAlchemy
psycopg2-binary
boto3
python-dotenv
slowapi
passlib